Initial release of ark_rcon.pl
This, of course, had lots and lots of prior history to it.
Clobbered all of it, however, as they all contained my password hard-coded as a
default.
Simply chaging the password would have fixed this, and of course I did that as
well, but I'd rather nobody saw that and tried to be all l33t h4x0r.

For the world at large, the prior history is probably meaningless anyway.